refactor: streamline and clean the dev/test/build pipeline
- streamline tsconfig and path alias
- bump dependencies
- remove unused code e.g. prod-paths, inexistent path aliases
- remove unused dependencies e.g. esbuild, markdown-it
- develop locally rather than in a docker wrapper
- simplify package.json scripts
- use .env variables when possible e.g. cogegen, apollo
- add a basic developer's guide
