+++ .env.example
@@ -5,7 +5,7 @@ POSTGRES_PASSWORD=secretpgpassword
 HASURA_GRAPHQL_DATABASE_URL=postgres://postgres:secretpgpassword@postgres:5432/postgres
 HASURA_GRAPHQL_JWT_SECRET='{"type":"HS256", "key":"5152fa850c02dc222631cca898ed1485821a70912a6e3649c49076912daa3b62182ba013315915d64f40cddfbb8b58eb5bd11ba225336a6af45bbae07ca873f3","issuer":"hasura-auth"}'
 HASURA_GRAPHQL_ADMIN_SECRET=hello123
-HASURA_GRAPHQL_GRAPHQL_URL=http://graphql-engine:8080/v1/graphql
+HASURA_GRAPHQL_GRAPHQL_URL=http://localhost:8080/v1/graphql
 
 # SERVER
 AUTH_HOST=
@@ -13,12 +13,12 @@ AUTH_PORT=4000
 AUTH_SERVER_URL=http://localhost:4000
 
 # SMTP
-AUTH_SMTP_HOST=mailhog
+AUTH_SMTP_HOST=localhost
 AUTH_SMTP_PORT=1025
 AUTH_SMTP_USER=user
 AUTH_SMTP_PASS=password
 AUTH_SMTP_SENDER=hasura-auth@example.com
-AUTH_SMTP_AUTH_METHOD
+AUTH_SMTP_AUTH_METHOD=
 AUTH_SMTP_SECURE=false
 
 # GRAVATAR
@@ -27,7 +27,7 @@ AUTH_GRAVATAR_DEFAULT=
 AUTH_GRAVATAR_RATING=
 
 # CLIENT & APP
-AUTH_CLIENT_URL=
+AUTH_CLIENT_URL=http://localhost:3000
 
 # SIGN UP
 AUTH_ANONYMOUS_USERS_ENABLED=false
