+++ .github/workflows/build.yml
@@ -120,6 +120,6 @@ jobs:
           password: ${{ secrets.DOCKER_PASSWORD }}
           repository: nhost/hasura-auth
           path: .
-          dockerfile: docker/prod/Dockerfile
+          dockerfile: Dockerfile
           tag_with_ref: true
           tag_with_sha: true
