+++ codegen.yaml
@@ -1,11 +1,14 @@
 overwrite: true
-schema: 'http://localhost:8080/v1/graphql'
+schema: '${HASURA_GRAPHQL_GRAPHQL_URL}'
 headers:
-  x-hasura-admin-secret: hello123
+  x-hasura-admin-secret: ${HASURA_GRAPHQL_ADMIN_SECRET}
 documents: 'src/gql/**/*.graphql'
 generates:
   src/utils/__generated__/graphql-request.ts:
     plugins:
+      - add:
+          content:
+            - /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unused-vars */
       - 'typescript'
       - 'typescript-operations'
       - 'typescript-graphql-request'
