--- docker/dev/docker-compose.yaml
@@ -1,114 +1,0 @@
-version: '3.6'
-services:
-  postgres:
-    image: nhost/postgres:12-v0.0.6
-    restart: always
-    volumes:
-      - ./db_data/db:/var/lib/postgresql/data
-    environment:
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-secretpgpassword}
-  graphql-engine:
-    image: hasura/graphql-engine:v2.1.0
-    depends_on:
-      - 'postgres'
-    restart: always
-    environment:
-      HASURA_GRAPHQL_DATABASE_URL: ${HASURA_GRAPHQL_DATABASE_URL}
-      HASURA_GRAPHQL_JWT_SECRET: ${HASURA_GRAPHQL_JWT_SECRET}
-      HASURA_GRAPHQL_ADMIN_SECRET: ${HASURA_GRAPHQL_ADMIN_SECRET}
-      HASURA_GRAPHQL_UNAUTHORIZED_ROLE: public
-      HASURA_GRAPHQL_LOG_LEVEL: debug
-      HASURA_GRAPHQL_ENABLE_CONSOLE: 'true'
-    ports:
-      - '8080:8080'
-  hasura-auth:
-    container_name: hasura-auth
-    image: nhost/hasura-auth:latest
-    build:
-      context: ../../
-      dockerfile: docker/dev/Dockerfile
-    restart: 'always'
-    depends_on:
-      - 'graphql-engine'
-      - 'mailhog'
-    ports:
-      - '4000:4000'
-    environment:
-      AUTH_HOST: ${AUTH_HOST:-0.0.0.0}
-      AUTH_PORT: ${AUTH_HOST:-4000}
-      AUTH_SERVER_URL: ${AUTH_SERVER_URL}
-
-      AUTH_SMTP_HOST: ${AUTH_SMTP_HOST}
-      AUTH_SMTP_PORT: ${AUTH_SMTP_PORT}
-      AUTH_SMTP_PASS: ${AUTH_SMTP_PASS}
-      AUTH_SMTP_USER: ${AUTH_SMTP_USER}
-      AUTH_SMTP_SECURE: ${AUTH_SMTP_SECURE}
-      AUTH_SMTP_SENDER: ${AUTH_SMTP_SENDER}
-
-      AUTH_SMS_PROVIDER: ${AUTH_SMS_PROVIDER}
-      AUTH_SMS_TWILIO_ACCOUNT_SID: ${AUTH_SMS_TWILIO_ACCOUNT_SID}
-      AUTH_SMS_TWILIO_AUTH_TOKEN: ${AUTH_SMS_TWILIO_AUTH_TOKEN}
-      AUTH_SMS_TWILIO_MESSAGING_SERVICE_ID: ${AUTH_SMS_TWILIO_MESSAGING_SERVICE_ID}
-
-      AUTH_CLIENT_URL: ${AUTH_CLIENT_URL:-http://localhost:3000}
-
-      AUTH_ANONYMOUS_USERS_ENABLED: ${AUTH_ANONYMOUS_USERS_ENABLED}
-      AUTH_DISABLE_NEW_USERS: ${AUTH_DISABLE_NEW_USERS}
-      AUTH_ACCESS_CONTROL_ALLOWED_EMAILS: ${AUTH_ACCESS_CONTROL_ALLOWED_EMAILS}
-      AUTH_ACCESS_CONTROL_ALLOWED_EMAIL_DOMAINS: ${AUTH_ACCESS_CONTROL_ALLOWED_EMAIL_DOMAINS}
-      AUTH_ACCESS_CONTROL_BLOCKED_EMAILS: ${AUTH_ACCESS_CONTROL_BLOCKED_EMAILS}
-      AUTH_ACCESS_CONTROL_BLOCKED_EMAIL_DOMAINS: ${AUTH_ACCESS_CONTROL_BLOCKED_EMAIL_DOMAINS}
-      AUTH_PASSWORD_MIN_LENGTH: ${AUTH_PASSWORD_MIN_LENGTH}
-      AUTH_PASSWORD_HIBP_ENABLED: ${AUTH_PASSWORD_HIBP_ENABLED}
-      AUTH_USER_DEFAULT_ROLE: ${AUTH_USER_DEFAULT_ROLE}
-      AUTH_USER_DEFAULT_ALLOWED_ROLES: ${AUTH_USER_DEFAULT_ALLOWED_ROLES}
-      AUTH_USER_ALLOWED_ROLES: ${AUTH_USER_ALLOWED_ROLES}
-      AUTH_LOCALE_DEFAULT: ${AUTH_LOCALE_DEFAULT}
-      AUTH_LOCALE_ALLOWED_LOCALES: ${AUTH_LOCALE_ALLOWED_LOCALES}
-
-      AUTH_EMAIL_PASSWORDLESS_ENABLED: ${AUTH_EMAIL_PASSWORDLESS_ENABLED}
-      AUTH_SMS_PASSWORDLESS_ENABLED: ${AUTH_SMS_PASSWORDLESS_ENABLED}
-      AUTH_PASSWORDLESS_SMS_TEMPLATE: ${AUTH_PASSWORDLESS_SMS_TEMPLATE}
-      AUTH_EMAIL_SIGNIN_EMAIL_VERIFIED_REQUIRED: ${AUTH_EMAIL_SIGNIN_EMAIL_VERIFIED_REQUIRED}
-      AUTH_ACCESS_CONTROL_ALLOWED_REDIRECT_URLS: ${AUTH_ACCESS_CONTROL_ALLOWED_REDIRECT_URLS}
-      AUTH_MFA_ENABLED: ${AUTH_MFA_ENABLED}
-      AUTH_MFA_TOTP_ISSUER: ${AUTH_MFA_TOTP_ISSUER}
-
-      AUTH_PROVIDER_GOOGLE_ENABLED: ${AUTH_PROVIDER_GOOGLE_ENABLED}
-      AUTH_PROVIDER_GOOGLE_CLIENT_ID: ${AUTH_PROVIDER_GOOGLE_CLIENT_ID}
-      AUTH_PROVIDER_GOOGLE_CLIENT_SECRET: ${AUTH_PROVIDER_GOOGLE_CLIENT_SECRET}
-      AUTH_PROVIDER_STRAVA_ENABLED: ${AUTH_PROVIDER_STRAVA_ENABLED}
-      AUTH_PROVIDER_STRAVA_CLIENT_ID: ${AUTH_PROVIDER_STRAVA_CLIENT_ID}
-      AUTH_PROVIDER_STRAVA_CLIENT_SECRET: ${AUTH_PROVIDER_STRAVA_CLIENT_SECRET}
-      AUTH_PROVIDER_GITHUB_ENABLED: ${AUTH_PROVIDER_GITHUB_ENABLED}
-      AUTH_PROVIDER_GITHUB_CLIENT_ID: ${AUTH_PROVIDER_GITHUB_CLIENT_ID}
-      AUTH_PROVIDER_GITHUB_CLIENT_SECRET: ${AUTH_PROVIDER_GITHUB_CLIENT_SECRET}
-
-      AUTH_LOG_LEVEL: ${AUTH_LOG_LEVEL}
-
-      HASURA_GRAPHQL_DATABASE_URL: ${HASURA_GRAPHQL_DATABASE_URL}
-      HASURA_GRAPHQL_JWT_SECRET: ${HASURA_GRAPHQL_JWT_SECRET}
-      HASURA_GRAPHQL_ADMIN_SECRET: ${HASURA_GRAPHQL_ADMIN_SECRET}
-      HASURA_GRAPHQL_GRAPHQL_URL: ${HASURA_GRAPHQL_GRAPHQL_URL:-http://graphql-engine:8080/v1/graphql}
-
-    volumes:
-      - '../../src:/app/src'
-      - '../../email-templates:/app/email-templates'
-      - '../../test:/app/test'
-      - '../../types:/app/types'
-      - '../../migrations:/app/migrations'
-      - '../../package.json:/app/package.json'
-      - '../../tsconfig.json:/app/tsconfig.json'
-      - '../../jest.config.js:/app/jest.config.js'
-  mailhog:
-    image: mailhog/mailhog
-    environment:
-      SMTP_HOST: mailhog
-      SMTP_PORT: 1025
-      SMTP_PASS: password
-      SMTP_USER: user
-      SMTP_SECURE: 'false'
-      SMTP_SENDER: hbp@hbp.com
-    ports:
-      - 1025:1025 # smtp server
-      - 8025:8025 # web ui
