+++ jest.config.js
@@ -1,13 +1,13 @@
 module.exports = {
   roots: ['./test'],
+  globals: {
+    server: null,
+  },
   globalSetup: '<rootDir>/test/global-setup.ts',
   verbose: false,
   moduleNameMapper: {
     '^@config/(.*)$': '<rootDir>/src/config/$1',
-    '^@middleware/(.*)$': '<rootDir>/src/middleware/$1',
-    '^@routes/(.*)$': '<rootDir>/src/routes/$1',
     '^@/(.*)$': '<rootDir>/src/$1',
-    '^test/(.*)$': '<rootDir>/test/$1',
   },
   testPathIgnorePatterns: [
     '<rootDir>/dist/',
