+++ package.json
@@ -8,23 +8,15 @@
   "version": "0.0.0",
   "main": "dist/start.js",
   "scripts": {
-    "test": "jest --runInBand",
-    "test:in-docker": "docker exec -it hasura-auth yarn test",
-    "dev:docker:up": "cd docker/dev; docker-compose up -d --build",
-    "dev:docker:logs": "cd docker/dev; docker-compose logs -f",
-    "dev:docker:down": "cd docker/dev; docker-compose down",
-    "dev:codegen": "graphql-codegen --config codegen.yaml",
-    "dev:in-docker": "node-dev -r tsconfig-paths/register --respawn --clear  --unhandled-rejections=strict  src/ts-start.ts",
-    "dev": "concurrently --restart-tries 5 --restart-after 5000 \"yarn dev:docker:up\" \"yarn dev:codegen -w\"",
-    "test:docker:up": "cd docker/test; docker-compose -f docker-compose.yaml up --build",
-    "test:docker:logs": "cd docker/test; docker-compose -f docker-compose.yaml logs -f",
-    "test:docker:down": "cd docker/test; docker-compose -f docker-compose.yaml down",
-    "test:watch": "concurrently --restart-tries 5 --restart-after 5000 \"yarn test:docker:up\" \"yarn dev:codegen -w\"",
-    "build": "tsc --project tsconfig.json",
-    "build:docker": "docker build -t nhost/hasura-auth:local -f docker/prod/Dockerfile .",
-    "start": "node -r ./dist/start.js",
+    "build": "rimraf dist && tsc --project tsconfig.build.json && tsc-alias",
+    "build:docker": "docker build -t nhost/hasura-auth:local .",
+    "start": "node ./dist/start.js",
     "lint": "eslint --ext .ts .",
-    "lint:fix": "npm run lint --fix"
+    "lint:fix": "npm run lint --fix",
+    "dev": "concurrently \"yarn dev:start\" \"yarn codegen --watch\"",
+    "codegen": "dotenv graphql-codegen --config codegen.yaml",
+    "dev:start": "ts-node-dev -r tsconfig-paths/register --respawn --clear  --unhandled-rejections=strict src/start.ts",
+    "test": "AUTH_LOG_LEVEL=error jest --runInBand"
   },
   "dependencies": {
     "@djgrant/postgres-migrations": "^4.1.0-rc.1",
@@ -42,16 +34,15 @@
     "express-boom": "^3.0.0",
     "express-joi-validation": "^5.0.0",
     "express-pino-logger": "^6.0.0",
-    "graphql": "15.0.0",
-    "graphql-request": "^3.4.0",
-    "graphql-tag": "2.11.0",
+    "graphql": "16.2.0",
+    "graphql-request": "^3.7.0",
+    "graphql-tag": "2.12.6",
     "gravatar": "^1.8.1",
     "helmet": "3.22.0",
     "hibp": "9.0.0",
     "joi": "^17.4.0",
     "jose": "1.28.1",
     "js-yaml": "3.13.1",
-    "module-alias": "^2.2.2",
     "nocache": "^2.1.0",
     "node-fetch": "^2.6.1",
     "nodemailer": "6.4.16",
@@ -74,10 +65,11 @@
     "uuid": "7.0.3"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "^1.21.7",
-    "@graphql-codegen/typescript": "^1.23.0",
-    "@graphql-codegen/typescript-graphql-request": "^3.2.5",
-    "@graphql-codegen/typescript-operations": "^1.18.4",
+    "@graphql-codegen/add": "^3.1.1",
+    "@graphql-codegen/cli": "^2.3.1",
+    "@graphql-codegen/typescript": "^2.4.2",
+    "@graphql-codegen/typescript-graphql-request": "^4.3.3",
+    "@graphql-codegen/typescript-operations": "^2.2.2",
     "@types/bcryptjs": "^2.4.2",
     "@types/body-parser": "1.19.0",
     "@types/cors": "2.8.6",
@@ -90,9 +82,9 @@
     "@types/gravatar": "^1.8.1",
     "@types/hapi__joi": "16.0.12",
     "@types/helmet": "0.0.45",
-    "@types/jest": "^26.0.23",
+    "@types/jest": "^27.4.0",
     "@types/lodash.kebabcase": "4.1.6",
-    "@types/node": "^13.11.1",
+    "@types/node": "^17.0.9",
     "@types/node-fetch": "^2.5.7",
     "@types/nodemailer": "6.4.0",
     "@types/passport": "1.0.3",
@@ -114,8 +106,7 @@
     "codecov": "^3.7.1",
     "concurrently": "^6.2.0",
     "cz-conventional-changelog": "3.1.0",
-    "dotenv-cli": "^4.0.0",
-    "esbuild": "^0.12.16",
+    "dotenv-cli": "^4.1.1",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.10.1",
     "eslint-plugin-jest": "23.8.2",
@@ -123,24 +114,20 @@
     "faker": "^5.5.3",
     "get-port": "^5.1.1",
     "husky": "4.2.5",
-    "jest": "^27.3.1",
+    "jest": "^27.4.7",
     "jest-extended": "0.11.5",
     "lint-staged": "10.1.3",
-    "markdown-it-multimd-table": "4.0.1",
-    "node-dev": "^7.0.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.2",
-    "pretty-quick": "2.0.1",
+    "rimraf": "^3.0.2",
     "supertest": "^6.1.6",
-    "ts-jest": "^27.0.3",
-    "ts-node": "^10.1.0",
-    "tsconfig-paths": "3.9.0",
-    "typescript": "^4.3.5"
+    "ts-jest": "^27.1.3",
+    "ts-node-dev": "^1.1.8",
+    "tsc-alias": "^1.5.0",
+    "tsconfig-paths": "3.12.0",
+    "typescript": "4.5.4"
   },
-  "_moduleAliases": {
-    "@middleware": "./dist/middleware",
-    "@routes": "./dist/routes",
-    "@config": "./dist/config",
-    "@": "./dist"
+  "engines": {
+    "node": ">=14.18.3"
   }
 }
