--- prod-paths.js
@@ -1,8 +1,0 @@
-const tsConfig = require('./tsconfig.json')
-const tsConfigPaths = require('tsconfig-paths')
-
-const baseUrl = './dist' // Either absolute or relative path. If relative it's resolved to current working directory.
-const cleanup = tsConfigPaths.register({
-  baseUrl,
-  paths: tsConfig.compilerOptions.paths
-})
