+++ src/helpers.ts
@@ -5,6 +5,8 @@ import bcrypt from 'bcryptjs';
 import { gqlSdk } from './utils/gqlSDK';
 // import { UserFieldsFragment } from './utils/__generated__/graphql-request';
 import { ENV } from './utils/env';
+import { logger } from './logger';
+import axios from 'axios';
 
 /**
  * Create QR code.
@@ -164,3 +166,33 @@ export function isValidRedirectTo({
 
   return false;
 }
+
+function delay(ms: number) {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
+
+const getHasuraReadyState = async () => {
+  try {
+    await axios.get(
+      `${ENV.HASURA_GRAPHQL_GRAPHQL_URL.replace('/v1/graphql', '/healthz')}`
+    );
+    return true;
+  } catch (err) {
+    return false;
+  }
+};
+
+export const waitForHasura = async () => {
+  let hasuraIsReady = false;
+
+  while (!hasuraIsReady) {
+    hasuraIsReady = await getHasuraReadyState();
+
+    if (hasuraIsReady) {
+      logger.info('Hasura is ready');
+    } else {
+      logger.info('Hasura is not ready. Retry in 5 seconds.');
+      await delay(5000);
+    }
+  }
+};
