+++ src/start.ts
@@ -1,2 +1,5 @@
-require('module-alias/register');
-import './ts-start';
+import 'dotenv/config';
+import './env-vars-check';
+import { start } from './server';
+
+start();
