+++ tsconfig.json
@@ -2,25 +2,20 @@
   "ts-node": {
     "files": true
   },
-  "exclude": ["node_modules", "**/*.test.ts", "test"],
+  "exclude": ["node_modules"],
   "include": ["src/**/*.ts", "test/**/*.ts", "types/**/*.ts"],
   "compilerOptions": {
     "lib": ["dom", "es2020"],
     "paths": {
-      // "*": ["types/*"],
-      "@config/*": ["src/config/*"],
-      "@middleware/*": ["src/middleware/*"],
-      "@routes/*": ["src/routes/*"],
-      "@/*": ["src/*"],
-      "test/*": ["test/*"],
-      "@test/*": ["test/*"]
+      "@config/*": ["config/*"],
+      "@/*": ["./*"]
     },
     "outDir": "dist",
     "noUnusedLocals": true,
     "module": "commonjs",
     "strict": true,
     "target": "es2019",
-    "baseUrl": ".",
+    "baseUrl": "src",
     "esModuleInterop": true,
     "moduleResolution": "node",
     "strictNullChecks": true,
